+++ lapce-data/src/data.rs
@@ -1730,29 +1730,6 @@ impl LapceTabData {
                     Cursor::new(CursorMode::Insert(Selection::caret(0)), None, None)
                 };
             }
-            LapceWorkbenchCommand::SourceControlCopyActiveFileRemoteUrl => {
-                if let Some(editor) = self.main_split.active_editor() {
-                    if let BufferContent::File(path) = &editor.content {
-                        let event_sink = ctx.get_external_handle();
-
-                        self.proxy.proxy_rpc.git_get_remote_file_url(
-                            path.clone(),
-                            move |result| {
-                                if let Ok(ProxyResponse::GitGetRemoteFileUrl {
-                                    file_url,
-                                }) = result
-                                {
-                                    let _ = event_sink.submit_command(
-                                        LAPCE_UI_COMMAND,
-                                        LapceUICommand::PutToClipboard(file_url),
-                                        Target::Auto,
-                                    );
-                                }
-                            },
-                        )
-                    }
-                }
-            }
             LapceWorkbenchCommand::SourceControlDiscardActiveFileChanges => {
                 if let Some(editor) = self.main_split.active_editor() {
                     if let BufferContent::File(path) = &editor.content {